+++ Panther/src/sema/SemanticAnalyzer.h
@@ -706,6 +706,7 @@ namespace pcit::panther{
 
 			struct InterfaceAttrs{
 				bool is_pub;
+				bool is_polymorphic;
 			};
 			EVO_NODISCARD auto analyze_interface_attrs(
 				const AST::InterfaceDef& interface_def,