+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Change Log
 
 <!---------------------------------->
+<a name="v0.0.180.0"></a>
+## v0.0.180.0
+
+### Panther
+- Added polymorphic/non-polymorphic interfaces
+- Fixed crashing or incorrect lowering when lowering use of return parameters
+- Fixed interface pointer accessors not computing mutablility properly
+
+
+<!---------------------------------->
+<a name="v0.0.179.2"></a>
+## v0.0.179.2
+
+### Panther
+- Fixed crashing when calling an interface method statement
+
+
+<!---------------------------------->
 <a name="v0.0.179.1"></a>
 ## v0.0.179.1